fix: Versioning of BPMN models #102

+++ connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/api/GenerateElementTemplate.java
@@ -45,5 +45,5 @@ public @interface GenerateElementTemplate {
      * is therefore not included in the element template.
      * @return the version of the element template
      */
-    double version();
+    int version();
 }
