+++ connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/impl/ElementTemplateGenerationResult.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 public class ElementTemplateGenerationResult {
 
     private String name;
-    private double version;
+    private int version;
     private String jsonString;
 }
