+++ connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/impl/ElementTemplateInfo.java
@@ -12,7 +12,7 @@ public class ElementTemplateInfo {
 
     private String name;
     private String id;
-    private double version;
+    private int version;
     private String type;
     private BPMNElementType[] appliesTo;
     private Class<?> inputType;
